+++ app/__init__.py
@@ -17,6 +17,9 @@ def create_app(config_object=Config) -> Flask:
     from app.api.atendente import bp as atendente_bp
     app.register_blueprint(atendente_bp, url_prefix="/api")
 
+    from app.api.servidores import bp as servidores_bp
+    app.register_blueprint(servidores_bp, url_prefix="/api")
+
     from app.pages import bp as pages_bp
     app.register_blueprint(pages_bp)